+++ angular2-token.js
@@ -2,4 +2,5 @@
 function __export(m) {
     for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
 }
-__export(require('./lib/angular2-token.service'));
+__export(require('./src/angular2-token.service'));
+__export(require('./src/a2t-ui/a2t-ui.module'));