+++ src/a2t-ui/a2t-shared/a2t-form/a2t-form.component.ts
@@ -1,0 +1,17 @@
+import { Component }  from '@angular/core';
+
+import {
+    A2tFormService,
+    BaseField,
+    A2tFormFieldComponent
+} from './';
+
+@Component({
+    selector: 'a2t-form',
+    templateUrl: 'a2t-form.component.html',
+    styleUrls: ['a2t-form.component.css']
+})
+
+export class A2tFormComponent {
+    constructor(private _formService: A2tFormService) { }
+}