+++ src/a2t-ui/a2t-shared/index.ts
@@ -1,0 +1,4 @@
+export * from './a2t-form/a2t-form.model';
+export * from './a2t-form/a2t-form.service';
+
+export * from './a2t-shared.module';