+++ src/a2t-ui/index.ts
@@ -1,0 +1,11 @@
+export * from './a2t-shared';
+
+export * from './a2t-sign-in/a2t-sign-in.component';
+export * from './a2t-sign-up/a2t-sign-up.component';
+export * from './a2t-reset-password/a2t-reset-password.component';
+export * from './a2t-update-password/a2t-update-password.component';
+
+export * from './a2t-ui.forms';
+export * from './a2t-ui.component';
+export * from './a2t-ui.routes';
+export * from './a2t-ui.module';